Clean up push; PG detection now in separate module, sym operations now in symtools
