+++ molsym/__init__.py
@@ -1,6 +1,5 @@
 from .molecule import Molecule
-from .flowchart import find_point_group
-#from .symtext.main import symtext_from_file, symtext_from_mol
+from .pgdetect.flowchart import find_point_group
 from .symtext.symtext import Symtext
 from .symmetrize import symmetrize
 from .import salcs
