+++ molsym/molecule.py
@@ -1,10 +1,9 @@
 import numpy as np
-#import psi4
 import qcelemental as qcel
-import json
 from dataclasses import dataclass
 from copy import deepcopy
 global_tol = 1e-6
+
 @dataclass
 class Atom():
     Z:int
@@ -30,25 +29,26 @@ class Molecule():
 
     @classmethod
     def from_schema(cls, schema):
-        atoms = schema["symbols"] # was symbols
+        atoms = schema["symbols"]
         natoms = len(atoms)
-        coords = np.reshape(schema["geometry"], (natoms,3)) # was geometry
+        coords = np.reshape(schema["geometry"], (natoms,3))
         # As of now, QCElemental seems to have issues assigning masses, so I do it
         masses = np.zeros(natoms)
         for (idx, symb) in enumerate(atoms):
             masses[idx] = qcel.periodictable.to_mass(symb)
-        #masses = schema["masses"]
         return cls(atoms, coords, masses)
 
     @classmethod
     def from_file(cls, fn):
         with open(fn, "r") as lfn:
             strang = lfn.read()
+        strang = "units bohr\n"+strang
         schema = qcel.models.Molecule.from_data(strang).dict()
         return cls.from_schema(schema)
 
     @classmethod
     def from_psi4_schema(cls, schema):
+        # Schemas coming from Psi4 are different for some reason?
         atoms = schema["elem"] # was symbols
         natoms = len(atoms)
         coords = np.reshape(schema["geom"], (natoms,3)) # was geometry
@@ -56,13 +56,13 @@ class Molecule():
         masses = np.zeros(natoms)
         for (idx, symb) in enumerate(atoms):
             masses[idx] = qcel.periodictable.to_mass(symb)
-        #masses = schema["masses"]
         return cls(atoms, coords, masses)
 
     def __repr__(self) -> str:
         rstr = "MolSym Molecule:\n"
         for i in range(self.natoms):
-            rstr += f"   {self.atoms[i]:3s}   {self.coords[i,0]:12.8f}   {self.coords[i,1]:12.8f}   {self.coords[i,2]:12.8f}\n"
+            rstr += f"   {self.atoms[i]:3s}   {self.coords[i,0]:12.8f}"
+            rstr_+= f"   {self.coords[i,1]:12.8f}   {self.coords[i,2]:12.8f}\n"
         return rstr
 
     def __str__(self) -> str:
@@ -81,7 +81,10 @@ class Molecule():
         else:
             eq_tol = other.tol
         if isinstance(other, Molecule):
-            return (other.atoms == self.atoms).all() and (other.masses == self.masses).all() and np.allclose(other.coords, self.coords, atol=eq_tol)
+            c1 = (other.atoms == self.atoms).all()
+            c2 = (other.masses == self.masses).all()
+            c3 = np.allclose(other.coords, self.coords, atol=eq_tol)
+            return c1 and c2 and c3
 
     def find_com(self):
         com = np.zeros(3)
@@ -89,19 +92,16 @@ class Molecule():
             com += self.masses[i]*self.coords[i,:]
         return com / sum(self.masses)
 
-    def translate(self, r):
-        for i in range(self.natoms):
-            self.coords[i,:] -= r
-        
     def is_at_com(self):
         if sum(abs(self.find_com())) < self.tol:
             return True
         else:
             return False
 
-    def _transform(self, M):
-        self.coords = np.dot(self.coords,np.transpose(M))
-
+    def translate(self, r):
+        for i in range(self.natoms):
+            self.coords[i,:] -= r
+        
     def transform(self, M):
         new_mol = deepcopy(self)
         new_mol.coords = np.dot(new_mol.coords,np.transpose(M))
@@ -137,17 +137,17 @@ class Molecule():
             if i in skip:
                 continue
             else:
-                biggun = [i]
+                collect = [i]
             
             for k in out:
                 if i in k:
                     if i == k[0]:
-                        biggun.append(k[1])
+                        collect.append(k[1])
                         skip.append(k[1])
                     else:
-                        biggun.append(k[0])
+                        collect.append(k[0])
                         skip.append(k[0])
-            SEAs.append(SEA("", biggun, np.zeros(3)))
+            SEAs.append(SEA("", collect, np.zeros(3)))
         return SEAs
 
     def symmetrize(self, asym_tol=0.05):
@@ -185,66 +185,3 @@ class Molecule():
             evalMat[i,i] = np.sqrt(evals[i])
         X = np.dot(evecs, evalMat)
         self.coords = X[:,-3:]
-
-def rotation_matrix(axis, theta):
-    cos_t = np.cos(theta)
-    sin_t = np.sin(theta)
-    # NOT NORMALIZING AXIS!!!
-    M = np.zeros((3,3))
-    M += 1 - cos_t
-    for i in range(3):
-        for j in range(3):
-            M[i,j] *= axis[i]*axis[j]
-    M += np.asarray([[cos_t, -axis[2]*sin_t, axis[1]*sin_t],[axis[2]*sin_t, cos_t, -axis[0]*sin_t],[-axis[1]*sin_t, axis[0]*sin_t, cos_t]])
-    return M
-
-def reflection_matrix(axis):
-    M = np.zeros((3,3))
-    for i in range(3):
-        for j in range(i,3):
-            if i == j:
-                M[i,i] = 1 - 2*(axis[i]**2)
-            else:
-                M[i,j] = -2 * axis[i] * axis[j]
-                M[j,i] = M[i,j]
-    return M
-
-def inversion_matrix():
-    return -1*np.identity(3)
-
-def Cn(axis, n):
-    theta = 2*np.pi/n
-    return rotation_matrix(axis, theta)
-
-def Sn(axis, n):
-    return np.dot(Cn(axis, n), reflection_matrix(axis))
-
-def isequivalent(A,B):
-    if A.tol >= B.tol:
-        eq_tol = A.tol
-    else:
-        eq_tol = B.tol
-    h = []
-    for i in range(A.natoms):
-        for j in range(B.natoms):
-            if A.masses[i] == B.masses[j]:
-                zs = abs(A.coords[i,:]-B.coords[j,:])
-                if np.allclose(zs, [0,0,0], atol=eq_tol):
-                    h.append(j)
-                    break
-    if len(h) == A.natoms:
-        return True
-    return False
-
-def calcmoit(atoms):
-    I = np.zeros((3,3))
-    atoms.translate(atoms.find_com())
-    for i in range(3):
-        for j in range(3):
-            if i == j:
-                for k in range(atoms.natoms):
-                    I[i,i] += atoms.masses[k]*(atoms.coords[k,(i+1)%3]**2+atoms.coords[k,(i+2)%3]**2)
-            else:
-                for k in range(atoms.natoms):
-                    I[i,j] -= atoms.masses[k]*atoms.coords[k,i]*atoms.coords[k,j]
-    return I
