+++ test/pgs/Cnh.py
@@ -1,7 +1,7 @@
 import numpy as np
 from numpy.linalg import matrix_power
 from molsym.symtext.symtext import Symel
-from molsym.molecule import Cn, reflection_matrix, inversion_matrix, Sn
+from molsym.symtools import Cn, reflection_matrix, inversion_matrix, Sn
 
 x = np.array([1,0,0])
 y = np.array([0,1,0])
