+++ test/test_symtext.py
@@ -164,12 +164,13 @@ order_test_set = [1,4,6,24,24]
 
 @pytest.mark.parametrize("i", [i for i in range(len(fns))])
 def test_Symtext(i):
+    angstrom_per_bohr = 0.529177249
     mol = molsym.Molecule.from_file("/home/smg13363/MolSym/test/xyz/"+fns[i]+".xyz")
     mol = molsym.symmetrize(mol)
     symtext = molsym.Symtext.from_molecule(mol)
     # Add mult table, symels, ctab, class map?
     assert (mol_test_set[i][0] == mol.atoms).all()
-    assert np.isclose(mol_test_set[i][1], mol.coords).all()
+    assert np.isclose(mol_test_set[i][1]*angstrom_per_bohr, mol.coords).all() # QCElemental performed undesired unit conv. in test set
     assert symtext.pg.str == pgs[i]
     assert (symtext.atom_map == atom_map_test_set[i]).all()
     assert complex_test_set[i] == symtext.complex
